requirements.txt
pandas>=1.4.2
requests>=2.27.1
pyyaml>=6.0
